requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Machine Learning & NLP
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Vector Databases & Search
faiss-cpu>=1.7.4
chromadb>=0.4.18
langchain>=0.1.0
langchain-community>=0.0.10
rank-bm25>=0.2.2

# Graph Databases & Network Analysis
networkx>=3.2
neo4j>=5.15.0

# Search Engines
elasticsearch>=8.11.0

# LLM Integrations
openai>=1.3.0
google-generativeai>=0.3.0
ollama>=0.1.7

# Web Scraping & Processing
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
scrapy>=2.11.0
markdownify>=0.11.6

# Data Processing
pyyaml>=6.0.1
aiofiles>=23.2.0
tqdm>=4.66.0

# Logging & Monitoring
loguru>=0.7.2

# CLI & Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: GPU Support (uncomment if using CUDA)
# torch>=2.1.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# faiss-gpu>=1.7.4

# Optional: Advanced Features
# jupyter>=1.0.0
# streamlit>=1.28.0
# gradio>=4.0.0
